test-requirements.txt
openupgradelib
# TODO these should be on PyPI
locomotivecms @ git+https://github.com/akretion/locomotivecms-python-client.git
# test requirements
requests_mock
vcrpy
vcrpy-unittest
unittest2 # For shopinvader test_controller, which inherits component
odoo-test-helper
httpx  # For FastAPI tests
odoo-addon-product-get-price-helper @ git+https://github.com/OCA/product-attribute@refs/pull/1445/head#subdirectory=setup/product_get_price_helper
odoo-addon-shopinvader-search-engine-image @ git+https://github.com/shopinvader/odoo-shopinvader@refs/pull/1453/head#subdirectory=setup/shopinvader_search_engine_image
